Naming changes from offline discussion.
